refactor: extract duplicated ROLE_PERMISSIONS map into shared api/lib/permissions.js
The ROLE_PERMISSIONS constant and getPermissionsForRoles() helper were
identically duplicated in login.js and google.js (~70 lines each).
Created api/lib/permissions.js as the single source of truth and
updated both auth handlers to import from it.

This eliminates the risk of permission drift between auth methods
where a permission change in one file but not the other could grant
different access levels depending on whether the user logs in via
email/password or Google OAuth.

Fixes #5089
