+++ api/lib/permissions.js
@@ -1,0 +1,104 @@
+const ROLE_PERMISSIONS = {
+  SUPER_ADMIN: [
+    "events:view",
+    "events:create",
+    "events:edit",
+    "events:delete",
+    "events:register",
+    "hackathons:view",
+    "hackathons:host",
+    "hackathons:participate",
+    "projects:view",
+    "projects:submit",
+    "projects:upvote",
+    "users:view",
+    "users:edit",
+    "users:delete",
+    "analytics:view",
+    "content:moderate",
+    "profile:edit",
+    "profile:view",
+    "notifications:manage",
+    "admin:access",
+  ],
+  ADMIN: [
+    "events:view",
+    "events:create",
+    "events:edit",
+    "events:delete",
+    "events:register",
+    "hackathons:view",
+    "hackathons:host",
+    "hackathons:participate",
+    "projects:view",
+    "projects:submit",
+    "projects:upvote",
+    "users:view",
+    "analytics:view",
+    "content:moderate",
+    "profile:edit",
+    "profile:view",
+    "notifications:manage",
+    "admin:access",
+  ],
+  ORGANIZER: [
+    "events:view",
+    "events:create",
+    "events:edit",
+    "events:register",
+    "hackathons:view",
+    "hackathons:host",
+    "hackathons:participate",
+    "projects:view",
+    "projects:submit",
+    "projects:upvote",
+    "analytics:view",
+    "profile:edit",
+    "profile:view",
+  ],
+  VOLUNTEER: [
+    "events:view",
+    "events:register",
+    "hackathons:view",
+    "hackathons:participate",
+    "projects:view",
+    "projects:submit",
+    "projects:upvote",
+    "content:moderate",
+    "profile:edit",
+    "profile:view",
+  ],
+  ATTENDEE: [
+    "events:view",
+    "events:register",
+    "hackathons:view",
+    "hackathons:participate",
+    "projects:view",
+    "projects:submit",
+    "projects:upvote",
+    "profile:edit",
+    "profile:view",
+  ],
+  USER: [
+    "events:view",
+    "events:register",
+    "projects:view",
+    "projects:submit",
+    "hackathons:view",
+    "hackathons:participate",
+    "profile:edit",
+    "profile:view",
+  ],
+};
+
+const getPermissionsForRoles = (roles) => {
+  const permissionsSet = new Set();
+  roles.forEach((role) => {
+    const normalizedRole = role.toUpperCase();
+    const perms = ROLE_PERMISSIONS[normalizedRole] || ROLE_PERMISSIONS.USER;
+    perms.forEach((perm) => permissionsSet.add(perm));
+  });
+  return Array.from(permissionsSet);
+};
+
+export { ROLE_PERMISSIONS, getPermissionsForRoles };
